move adminhtml js, create new group of tables for customer addresses, add multiline field type for form, move subscription logic to newsletter module, fix address field orders for existing customer

+++ app/code/core/Mage/Eav/Model/Entity/Setup.php
@@ -352,16 +352,19 @@ class Mage_Eav_Model_Entity_Setup extends Mage_Core_Model_Resource_Setup
             $id = $this->getAttribute($entityTypeId, $id, 'attribute_id');
         }
         if (!is_numeric($id)) {
-            throw Mage::exception('Mage_Eav', __('Wrong attribute ID'));
+            //throw Mage::exception('Mage_Eav', __('Wrong attribute ID'));
+            return false;
         }
         return $id;
     }
 
     public function removeAttribute($entityTypeId, $code)
     {
-        $this->_conn->delete($this->getTable('eav/attribute'),
-            $this->_conn->quoteInto('attribute_id=?', $this->getAttributeId($entityTypeId, $code))
-        );
+        if ($attributeId = $this->getAttributeId($entityTypeId, $code)) {
+            $this->_conn->delete($this->getTable('eav/attribute'),
+                $this->_conn->quoteInto('attribute_id=?', $attributeId)
+            );
+        }
         return $this;
     }
